chore: organize directory structure
