+++ src/actions/action.js
@@ -1,0 +1,4 @@
+const PUT = 'cart/PUT';
+const INITIALIZE = 'productList/INITIALIZE';
+
+export { PUT, INITIALIZE };
