+++ src/components/productItem/ProductItem.jsx
@@ -1,49 +1,22 @@
 import { useState, useRef } from 'react';
 import { useSelector } from 'react-redux';
-import styled, { css } from 'styled-components';
-import Image from 'components/shared/image/Image';
-import ShoppingCartIcon from 'components/shared/icon/ShoppingCartIcon';
-import Modal from 'components/shared/modal/Modal';
-import { PUT } from 'modules/reducer';
-import store from 'app/store';
-import Button from 'components/shared/button/Button';
 
+import Image from 'components/shared/image/Image';
+import CartIcon from 'components/shared/cartIcon/CartIcon';
+import Modal from 'components/shared/modal/Modal';
+import Button from 'components/shared/button/Button';
+import Text from 'components/shared/text/Text';
 import { ReactComponent as PlusIcon } from 'assets/plus_icon.svg';
 import { ReactComponent as MinusIcon } from 'assets/minus_icon.svg';
-import Text from './shared/text/Text';
 
-const StyledProductItem = styled.div`
-  width: 282px;
-  position: relative;
-`;
+import store from 'store/store';
+import { PUT } from 'actions/action';
 
-const StyledProductContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 75px;
-  padding: 0 15px;
-`;
-
-const StyledProductText = styled.p`
-  font-weight: 400;
-  letter-spacing: 0.5px;
-  color: #333333;
-
-  ${props =>
-    props.name &&
-    css`
-      font-size: 16px;
-      line-height: 22px;
-    `}
-
-  ${props =>
-    props.price &&
-    css`
-      font-size: 20px;
-      line-height: 27px;
-    `}
-`;
+import {
+  StyledProductItem,
+  StyledProductContainer,
+  StyledProductText,
+} from 'components/productItem/style';
 
 const ProductItem = ({ id }) => {
   const [isOpen, setIsOpen] = useState(false);
@@ -129,7 +102,7 @@ const ProductItem = ({ id }) => {
               </div>
             </Button>
           ) : (
-            <ShoppingCartIcon onClick={handleClick} />
+            <CartIcon onClick={handleClick} />
           )}
         </div>
       </StyledProductContainer>
