+++ src/components/shared/cartIcon/CartIcon.jsx
@@ -1,0 +1,7 @@
+import StyledCartIcon from 'components/shared/cartIcon/style';
+
+const CartIcon = ({ ...rest }) => {
+  return <StyledCartIcon {...rest} />;
+};
+
+export default CartIcon;
