+++ src/components/shared/image/Image.jsx
@@ -1,9 +1,4 @@
-import styled from 'styled-components';
-
-const StyledImage = styled.img`
-  width: ${props => props.size || '282px'};
-  height: ${props => props.size || '282px'};
-`;
+import StyledImage from 'components/shared/image/style';
 
 const Image = ({ ...rest }) => {
   return <StyledImage {...rest} />;
