+++ src/components/shared/image/style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const StyledImage = styled.img`
+  width: ${props => props.size || '282px'};
+  height: ${props => props.size || '282px'};
+`;
+
+export default StyledImage;
