+++ src/components/shared/modal/Modal.jsx
@@ -1,18 +1,4 @@
-import styled from 'styled-components';
-
-const StyledModal = styled.div`
-  position: absolute;
-  display: flex;
-  border-radius: 25px;
-  width: 180px;
-  height: 50px;
-  bottom: 90px;
-  left: 45px;
-  background-color: white;
-  box-shadow: 0px 0px 4px rgb(0 0 0 / 20%);
-  align-items: center;
-  justify-content: space-around;
-`;
+import StyledModal from 'components/shared/modal/style';
 
 const Modal = ({ children, ...rest }) => {
   return <StyledModal {...rest}>{children}</StyledModal>;
