+++ src/components/shared/text/style.js
@@ -1,0 +1,7 @@
+import StyledText from 'components/shared/text/style';
+
+const Text = ({ children, ...rest }) => {
+  return <StyledText {...rest}>{children}</StyledText>;
+};
+
+export default Text;
